requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
fastapi==0.71.0
h11==0.12.0
idna==3.3
opensea-api==0.1.2
pydantic==1.9.0
requests==2.26.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
